requirements.txt
Flask
tensorflow
torch
torchvision
Pillow
numpy
requests
gunicorn
